pyproject.toml
[tool.poetry]
name = "svdi"
version = "1.0.0"
description = "Image compression and decompression tool using SVD"
authors = ["Vadim Yakshigulov <toadharvard@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
typer = {extras = ["all"], version = "^0.12.3"}
numpy = "^1.26.4"
pillow = "^10.3.0"
scipy = "^1.13.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
svdi = "svdi.cli:app"


[tool.ruff]
lint.ignore = ["E731"]

